+++ Changes.md
@@ -1,6 +1,11 @@
 0.61.x.x (relative to 0.61.5.0)
 ========
 
+Improvements
+------------
+
+- Spreadsheet : Improved drag and drop of lists of strings. <kbd>Shift</kbd> may now be held to add to the current list, and <kbd>Control</kbd> may be held to remove from the current list (#3894). This is particularly useful when a PathFilter has been added to a spreadsheet.
+
 Fixes
 -----